requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.28.0
pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.12.0
gradio>=3.32.0
ftfy>=6.1.1
regex>=2022.10.31
pycocotools>=2.0.6
accelerate>=0.19.0
deepspeed>=0.9.0
nltk>=3.8
rouge>=1.0.1
bitsandbytes>=0.39.0
peft>=0.4.0
safetensors>=0.3.1
sentencepiece>=0.1.99
protobuf>=3.20.0
optimum>=1.8.8